Rename rx -> rx2
