+++ app-example/src/main/java/com/badoo/ribs/example/feed_container/FeedContainer.kt
@@ -1,6 +1,6 @@
 package com.badoo.ribs.example.feed_container
 
-import com.badoo.ribs.rx.clienthelper.connector.Connectable
+import com.badoo.ribs.rx2.clienthelper.connector.Connectable
 import com.badoo.ribs.core.Rib
 import com.badoo.ribs.core.customisation.RibCustomisation
 import com.badoo.ribs.example.auth.AuthDataSource
