+++ app-example/src/main/java/com/badoo/ribs/example/photo_feed/PhotoFeedNode.kt
@@ -1,7 +1,7 @@
 package com.badoo.ribs.example.photo_feed
 
-import com.badoo.ribs.rx.clienthelper.connector.Connectable
-import com.badoo.ribs.rx.clienthelper.connector.NodeConnector
+import com.badoo.ribs.rx2.clienthelper.connector.Connectable
+import com.badoo.ribs.rx2.clienthelper.connector.NodeConnector
 import com.badoo.ribs.core.Node
 import com.badoo.ribs.core.modality.BuildParams
 import com.badoo.ribs.core.plugin.Plugin
