+++ libraries/rib-portal-rx/src/main/java/com/badoo/ribs/portal/RxPortalNode.kt
@@ -8,7 +8,7 @@ import com.badoo.ribs.portal.PortalRouter.Configuration
 import com.badoo.ribs.portal.PortalRouter.Configuration.Content
 import com.badoo.ribs.routing.source.backstack.BackStack
 import com.badoo.ribs.routing.source.backstack.operation.push
-import com.badoo.ribs.rx.workflows.RxWorkflowNode
+import com.badoo.ribs.rx2.workflows.RxWorkflowNode
 import io.reactivex.Single
 
 internal class RxPortalNode(
