+++ libraries/rib-rx/src/main/java/com/badoo/ribs/rx2/Disposables.kt
@@ -1,4 +1,4 @@
-package com.badoo.ribs.rx
+package com.badoo.ribs.rx2
 
 import com.badoo.ribs.core.plugin.NodeLifecycleAware
 import com.badoo.ribs.core.plugin.Plugin
