+++ sandbox/src/main/java/com/badoo/ribs/sandbox/rib/blocker/Blocker.kt
@@ -1,6 +1,6 @@
 package com.badoo.ribs.sandbox.rib.blocker
 
-import com.badoo.ribs.rx.clienthelper.connector.Connectable
+import com.badoo.ribs.rx2.clienthelper.connector.Connectable
 import com.badoo.ribs.core.Rib
 import com.badoo.ribs.core.customisation.RibCustomisation
 import com.badoo.ribs.sandbox.rib.blocker.Blocker.Input
