+++ sandbox/src/main/java/com/badoo/ribs/sandbox/rib/hello_world/HelloWorldNode.kt
@@ -1,14 +1,14 @@
 package com.badoo.ribs.sandbox.rib.hello_world
 
 import android.util.Log
-import com.badoo.ribs.rx.clienthelper.connector.Connectable
-import com.badoo.ribs.rx.clienthelper.connector.NodeConnector
+import com.badoo.ribs.rx2.clienthelper.connector.Connectable
+import com.badoo.ribs.rx2.clienthelper.connector.NodeConnector
 import com.badoo.ribs.core.modality.BuildParams
 import com.badoo.ribs.core.plugin.Plugin
 import com.badoo.ribs.core.view.RibView
 import com.badoo.ribs.sandbox.rib.hello_world.HelloWorld.Input
 import com.badoo.ribs.sandbox.rib.hello_world.HelloWorld.Output
-import com.badoo.ribs.rx.workflows.RxWorkflowNode
+import com.badoo.ribs.rx2.workflows.RxWorkflowNode
 import io.reactivex.Single
 
 class HelloWorldNode(
