+++ templates/snapshot/src/main/java/com/badoo/ribs/template/leaf/foo_bar/FooBarNode.kt
@@ -1,13 +1,13 @@
 package com.badoo.ribs.template.leaf.foo_bar
 
-import com.badoo.ribs.rx.clienthelper.connector.Connectable
-import com.badoo.ribs.rx.clienthelper.connector.NodeConnector
+import com.badoo.ribs.rx2.clienthelper.connector.Connectable
+import com.badoo.ribs.rx2.clienthelper.connector.NodeConnector
 import com.badoo.ribs.core.modality.BuildParams
 import com.badoo.ribs.core.plugin.Plugin
 import com.badoo.ribs.core.view.RibView
 import com.badoo.ribs.template.leaf.foo_bar.FooBar.Input
 import com.badoo.ribs.template.leaf.foo_bar.FooBar.Output
-import com.badoo.ribs.rx.workflows.RxWorkflowNode
+import com.badoo.ribs.rx2.workflows.RxWorkflowNode
 import io.reactivex.Single
 
 class FooBarNode internal constructor(
