+++ templates/snapshot/src/main/java/com/badoo/ribs/template/node_dagger/foo_bar/builder/FooBarModule.kt
@@ -3,7 +3,7 @@ package com.badoo.ribs.template.node_dagger.foo_bar.builder
 
 import com.badoo.ribs.core.modality.BuildParams
 import com.badoo.ribs.routing.source.backstack.BackStack
-import com.badoo.ribs.rx.disposables
+import com.badoo.ribs.rx2.disposables
 import com.badoo.ribs.template.node_dagger.foo_bar.FooBar
 import com.badoo.ribs.template.node_dagger.foo_bar.FooBarInteractor
 import com.badoo.ribs.template.node_dagger.foo_bar.FooBarNode
